+++ internal/config/config.go
@@ -55,7 +55,7 @@ type LoggingConfig struct {
 // IBEConfig holds Identity-Based Encryption configuration
 type IBEConfig struct {
 	DomainKeysDir string // Directory containing domain-specific master keys
-	KeyVersion    int    // Current key version
+	KeyVersion    int32  // Current key version
 	Salt          string // Salt for fingerprint generation (defaults to "fingerprint_salt_v1")
 	KeyRotation   struct {
 		Enabled     bool
@@ -164,7 +164,7 @@ func Load() (*Config, error) {
 		},
 		IBE: IBEConfig{
 			DomainKeysDir: getEnv("IBE_DOMAIN_KEYS_DIR", "./keys/domains"),
-			KeyVersion:    getEnvAsInt("IBE_KEY_VERSION", 1),
+			KeyVersion:    int32(getEnvAsInt("IBE_KEY_VERSION", 1)),
 			Salt:          getEnv("IBE_SALT", "fingerprint_salt_v1"),
 			KeyRotation: struct {
 				Enabled     bool
